Check that at least one document was selected

+++ main.py
@@ -3,6 +3,9 @@ class BulkDelete(SoftTimeOutAddOn):
     def main(self):
         confirm = self.data.get("confirm")
         if confirm is not None and confirm is True:
+            if not self.documents:
+                self.set_message("Please select at least one document.")
+                return
             for document in self.get_documents():
                 document.delete()
         else:
